Add TQ1 Segment
The TQ1 segment is used to specify the complex timing of events and
actions such as those that occur in order management and scheduling
systems. This segment determines the quantity, frequency, priority and
timing of a service.  By allowing the segment to repeat, it is possible to
have service requests that vary the quantity, frequency and priority of a
service request over time.