Add RXR Segment
4.14.2RXR - Pharmacy/Treatment Route Segment
The Pharmacy/Treatment Route segment contains the alternative combination
of route, site, administration device, and administration method that are
prescribed as they apply to a particular order.  The pharmacy, treatment
staff and/or nursing staff has a choice between the routes based on either
their professional judgment or administration instructions provided by the
physician.